+++ volmdlr/faces.py
@@ -58,13 +58,13 @@ class Face3D(volmdlr.core.Primitive3D):
                  and self.surface2d == other_.surface2d)
         return equal
 
-    def point_belongs(self, point3d: volmdlr.Point3D):
+    def point_belongs(self, point3d: volmdlr.Point3D, tol: float = 1e-6):
         """
         Tells you if a point is on the 3D face and inside its contour.
         """
         point2d = self.surface3d.point3d_to_2d(point3d)
         check_point3d = self.surface3d.point2d_to_3d(point2d)
-        if check_point3d.point_distance(point3d) > 1e-6:
+        if check_point3d.point_distance(point3d) > tol:
             return False
 
         return self.surface2d.point_belongs(point2d)
@@ -1118,6 +1118,18 @@ class Face3D(volmdlr.core.Primitive3D):
                 inner_contours_connected_cutting_contour[cutting_contour].append(new_inner_contour)
         return (dict_cutting_contour_intersections, dict_inner_contour_intersections,
                 inner_contours_connected_cutting_contour, list_cutting_contours)
+
+    def is_linesegment_crossing(self, linesegment):
+        """Verify if a face 3d is being crossed by a linesegment 3d. """
+        if not self.bounding_box.bbox_intersection(linesegment.bounding_box):
+            return False
+        triangulation = self.triangulation()
+        faces_triangulation = triangulation.triangular_faces()
+        for face in faces_triangulation:
+            inters = face.linesegment_intersections(linesegment)
+            if inters:
+                return True
+        return False
 
 
 class PlaneFace3D(Face3D):
@@ -3146,3 +3158,15 @@ class BSplineFace3D(Face3D):
             inner_contours=[plane3d.contour3d_to_2d(contour) for contour in self.inner_contours3d])
 
         return PlaneFace3D(surface3d=plane3d, surface2d=surface2d)
+
+    def linesegment_intersections(self, linesegment: vme.LineSegment3D) -> List[volmdlr.Point3D]:
+        """Aproximation of intersections between a bspline face 3D and a line segment 3D."""
+        triangulation = self.triangulation()
+        faces_triangulation = triangulation.triangular_faces()
+        linesegment_intersections = []
+        for face in faces_triangulation:
+            inters = face.linesegment_intersections(linesegment)
+            for point in inters:
+                if volmdlr.core.point_in_list(point, linesegment_intersections):
+                    linesegment_intersections.append(point)
+        return linesegment_intersections